move verification logic to signature classes
