+++ src/NuGet.Core/NuGet.Packaging/Signing/Timestamp/Timestamp.cs
@@ -3,6 +3,10 @@
 
 using System;
 using NuGet.Common;
+using System.Collections.Generic;
+using System.Globalization;
+using System.Security.Cryptography.X509Certificates;
+using System.Linq;
 
 #if IS_DESKTOP
 using System.Security.Cryptography.Pkcs;
@@ -84,6 +88,111 @@ namespace NuGet.Packaging.Signing
             {
                 throw new TimestampException(NuGetLogCode.NU3021, Strings.TimestampSignatureValidationFailed);
             }
+        }
+
+        /// <summary>
+        /// 
+        /// </summary>
+        /// <param name="signature"></param>
+        /// <param name="allowIgnoreTimestamp"></param>
+        /// <param name="allowUnknownRevocation"></param>
+        /// <param name="issues"></param>
+        /// <returns></returns>
+        internal bool Verify(
+            Signature signature,
+            bool allowIgnoreTimestamp,
+            bool allowUnknownRevocation,
+            HashAlgorithmName fingerprintAlgorithm,
+            List<SignatureLog> issues)
+        {
+            var treatIssueAsError = !allowIgnoreTimestamp;
+            var timestamperCertificate = SignerInfo.Certificate;
+            if (timestamperCertificate == null)
+            {
+                issues?.Add(SignatureLog.Issue(treatIssueAsError, NuGetLogCode.NU3020, Strings.TimestampNoCertificate));
+                return false;
+            }
+
+            if (VerificationUtility.IsTimestampValid(this, signature, treatIssueAsError, issues, SigningSpecifications.V1))
+            {
+                issues?.Add(SignatureLog.InformationLog(string.Format(CultureInfo.CurrentCulture, Strings.TimestampValue, GeneralizedTime.LocalDateTime.ToString()) + Environment.NewLine));
+
+                issues?.Add(SignatureLog.InformationLog(string.Format(CultureInfo.CurrentCulture,
+                    Strings.VerificationTimestamperCertDisplay,
+                    $"{Environment.NewLine}{CertificateUtility.X509Certificate2ToString(timestamperCertificate, fingerprintAlgorithm)}")));
+
+                var certificateExtraStore = SignedCms.Certificates;
+
+                using (var chainHolder = new X509ChainHolder())
+                {
+                    var chain = chainHolder.Chain;
+
+                    // This flags should only be set for verification scenarios, not signing
+                    chain.ChainPolicy.VerificationFlags = X509VerificationFlags.IgnoreNotTimeValid | X509VerificationFlags.IgnoreCtlNotTimeValid;
+
+                    CertificateChainUtility.SetCertBuildChainPolicy(chain.ChainPolicy, certificateExtraStore, DateTime.Now, NuGetVerificationCertificateType.Timestamp);
+
+                    var chainBuildSucceed = CertificateChainUtility.BuildCertificateChain(chain, timestamperCertificate, out var chainStatusList);
+
+                    issues?.Add(SignatureLog.DetailedLog(CertificateUtility.X509ChainToString(chain, fingerprintAlgorithm)));
+
+                    if (chainBuildSucceed)
+                    {
+                        return true;
+                    }
+
+                    var chainBuildingHasIssues = false;
+                    IEnumerable<string> messages;
+
+                    var timestampInvalidCertificateFlags = CertificateChainUtility.DefaultObservedStatusFlags |
+                        (X509ChainStatusFlags.Revoked) |
+                        (X509ChainStatusFlags.NotTimeValid) |
+                        (X509ChainStatusFlags.CtlNotTimeValid);
+
+                    if (CertificateChainUtility.TryGetStatusMessage(chainStatusList, timestampInvalidCertificateFlags, out messages))
+                    {
+                        foreach (var message in messages)
+                        {
+                            issues?.Add(SignatureLog.Issue(treatIssueAsError, NuGetLogCode.NU3028, message));
+                        }
+
+                        chainBuildingHasIssues = true;
+                    }
+
+                    // For all the special cases, chain status list only has unique elements for each chain status flag present
+                    // therefore if we are checking for one specific chain status we can use the first of the returned list
+                    // if we are combining checks for more than one, then we have to use the whole list.
+
+                    const X509ChainStatusFlags RevocationStatusFlags = X509ChainStatusFlags.RevocationStatusUnknown | X509ChainStatusFlags.OfflineRevocation;
+                    if (CertificateChainUtility.TryGetStatusMessage(chainStatusList, RevocationStatusFlags, out messages))
+                    {
+                        if (treatIssueAsError)
+                        {
+                            foreach (var message in messages)
+                            {
+                                issues?.Add(SignatureLog.Issue(!allowUnknownRevocation, NuGetLogCode.NU3028, message));
+                            }
+                        }
+
+                        if (!chainBuildingHasIssues && (allowIgnoreTimestamp || allowUnknownRevocation))
+                        {
+                            return true;
+                        }
+
+                        chainBuildingHasIssues = true;
+                    }
+
+                    // Debug log any errors
+                    issues?.Add(
+                        SignatureLog.DebugLog(
+                            string.Format(
+                                CultureInfo.CurrentCulture,
+                                Strings.ErrorInvalidCertificateChain,
+                                string.Join(", ", chainStatusList.Select(x => x.Status.ToString())))));
+                }
+            }
+
+            return false;
         }
 #endif
     }
